% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origene_primer_search.R
\name{origene_primer_search}
\alias{origene_primer_search}
\title{origene_primer_search}
\usage{
origene_primer_search(
  gene = "ITGA9",
  speice = "Human",
  ncbi_id = NULL,
  use_ncbi = F
)
}
\arguments{
\item{gene}{str gene smbol}

\item{speice}{str default as Human}

\item{ncbi_id}{the ncbi id of gene,default as NULL}

\item{use_ncbi}{use ncbi database when True,default as False}
}
\value{
vector
}
\description{
origene_primer_search for primer design
}
\examples{
origene_primer_search(gene = "itga9",
                                      speice = "Human",
                                      use_ncbi = F)
}
